zpodsdk/pyproject.toml
[tool.poetry]
name = "zpodsdk"
version = "0.3.0"
description = "A client library for accessing zPod API"

authors = []

readme = "README.md"
packages = [
    {include = "zpodsdk", from = "src"},
]
include = ["CHANGELOG.md", "zpod/py.typed"]

[tool.poetry.dependencies]
python = "^3.8.1"
httpx = "0.27.0"
attrs = "23.2.0"
python-dateutil = "2.9.0.post0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
